Move pkg/kafka to internal/kafka (#6908)
- part of https://github.com/jaegertracing/jaeger/issues/6869

- moved `pkg/kafka` to `internal/kafka`

-

- [ ] I have read
https://github.com/jaegertracing/jaeger/blob/master/CONTRIBUTING_GUIDELINES.md
- [ ] I have signed all commits
- [ ] I have added unit tests for the new functionality
- [ ] I have run lint and test steps successfully
  - for `jaeger`: `make lint test`
  - for `jaeger-ui`: `npm run lint` and `npm run test`
